GOB: move init code into seperate folder for init
move all init related source code files into an folder for better code overview, like for save or sound
