+++ engines/gob/gob.cpp
@@ -44,7 +44,7 @@
 #include "gob/dataio.h"
 #include "gob/game.h"
 #include "gob/sound/sound.h"
-#include "gob/init.h"
+#include "gob/init/init.h"
 #include "gob/inter/inter.h"
 #include "gob/draw/draw.h"
 #include "gob/goblin/goblin.h"
