+++ engines/gob/init/init_v4.cpp
@@ -26,33 +26,22 @@
  */
 
 #include "common/endian.h"
+#include "common/config-manager.h"
 
 #include "gob/gob.h"
-#include "gob/init.h"
-#include "gob/game.h"
+#include "gob/init/init.h"
 #include "gob/global.h"
 
 namespace Gob {
 
-Init_v3::Init_v3(GobEngine *vm) : Init_v2(vm) {
+Init_v4::Init_v4(GobEngine *vm) : Init_v3(vm) {
 }
 
-Init_v3::~Init_v3() {
+Init_v4::~Init_v4() {
 }
 
-void Init_v3::updateConfig() {
-// In the CD and Windows version of Goblins3, some texts are flagged
-// 'subtitles' incorrectly and therefore should be displayed in all cases.
-// The first obvious example is just after Death level.
-	if ((_vm->getGameType() == kGameTypeGob3) &&
-			(_vm->isCD() || (_vm->getPlatform() == Common::kPlatformWindows)))
-		_vm->_global->_doSubtitles = true;
-}
-
-void Init_v3::initVideo() {
-	Init_v2::initVideo();
-
-	_vm->_game->_noScroll = false;
+void Init_v4::updateConfig() {
+	_vm->_global->_doSubtitles = ConfMan.getBool("subtitles");
 }
 
 } // End of namespace Gob
